refactor(extensions): move app tool modules into a tools/ package
Groups uiux_tools/image_tools/video_tools/map_tools/browser_tools under
openjiuwen/extensions/app/tools/ instead of the extension's top-level
namespace, and updates imports, tests, and the README to match.
