+++ openjiuwen/extensions/app/agent.py
@@ -9,9 +9,9 @@ from openjiuwen.core.single_agent.schema.agent_card import AgentCard
 from openjiuwen.harness.tools import WebFetchWebpageTool, WebFreeSearchTool
 
 from . import config as app_config
-from .browser_tools import browser_inspect_page
 from .rails import A2uiToolEventRail
-from .uiux_tools import ALL_TOOLS
+from .tools.browser_tools import browser_inspect_page
+from .tools.uiux_tools import ALL_TOOLS
 
 AGENT_ID = "a2ui_react_agent"
