+++ openjiuwen/extensions/app/tools/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: UTF-8 -*-
+# Copyright (c) Huawei Technologies Co., Ltd. 2025. All rights reserved.
+"""Tools the ReAct agent registers -- see ``uiux_tools.ALL_TOOLS``, which
+assembles every tool from this package for ``agent.py`` to register.
+"""
